requirements.txt
# Core
Django==6.0.1
djangorestframework
django-cors-headers

# Auth
django-allauth
djangorestframework-simplejwt

# Database
psycopg2-binary
dj-database-url

# Cache / Queue
redis
django-redis
celery

# Production server & static files
gunicorn
whitenoise

requests

# Email (Step 6)
resend
sendgrid
django-sendgrid-v5

# QR Codes
qrcode[pil]

python-dotenv

# Billing
stripe